Update Transform/Task to use IgniteAgenda for consistency
Also some minor dependency updates.

This does require that we look up the AgendaTemplate twice, as the `hack` to create execProgress needs the template.  This will eventually go away which should simplify the logic.

Started to stub the isResubmit functionality, but some todos.  It should be still in functionally the same state as it was before, so we can verify working, then move onto implementing the resubmit logic

[git-p4: depot-paths = "//depot/service/java/projects/DistributedFileHandling/ContentProcessing/endpoint/modules/BaseEndpoint/main/": change = 540902]